Rename retry/repeat common file
