+++ Rx/v2/src/rxcpp/operators/rx-retry.hpp
@@ -21,7 +21,7 @@
 #define RXCPP_OPERATORS_RX_RETRY_HPP
 
 #include "../rx-includes.hpp"
-#include "impl-retry-repeat.hpp"
+#include "rx-retry-repeat-common.hpp"
 
 namespace rxcpp {
